+++ e3nn_jax/_src/graph_util_test.py
@@ -3,17 +3,22 @@ import jax.numpy as jnp
 import numpy as np
 
 
-def test_index_add():
+def test_scatter_sum():
     i = jnp.array([0, 2, 2, 0])
     x = jnp.array([1.0, 2.0, 3.0, -10.0])
     np.testing.assert_allclose(
-        e3nn.index_add(i, x, out_dim=3),
+        e3nn.scatter_sum(x, dst=i, output_size=3),
         jnp.array([-9.0, 0.0, 5.0]),
     )
 
     np.testing.assert_allclose(  # map_back
-        e3nn.index_add(i, x, map_back=True),
+        e3nn.scatter_sum(x, dst=i, map_back=True),
         jnp.array([-9.0, 5.0, 5.0, -9.0]),
+    )
+
+    np.testing.assert_allclose(  # nel
+        e3nn.scatter_sum(jnp.array([1.0, 2.0, 1.0, 0.5, 0.5, 0.7, 0.2, 0.1]), nel=jnp.array([3, 2, 3])),
+        jnp.array([4.0, 1.0, 1.0]),
     )